Some more semantic fixes

+++ QGLViewer/camera.cpp
@@ -964,8 +964,8 @@ Vec Camera::pointUnderPixel(const QPoint &pixel, bool &found) const {
   // Qt uses upper corner for its origin while GL uses the lower corner.
   glReadPixels(pixel.x(), screenHeight() - 1 - pixel.y(), 1, 1,
                GL_DEPTH_COMPONENT, GL_FLOAT, &depth);
-  found = depth < 1.0;
-  Vec point(pixel.x(), pixel.y(), depth);
+  found = static_cast<double>(depth)< 1.0;
+  Vec point(pixel.x(), pixel.y(), static_cast<double>(depth));
   point = unprojectedCoordinatesOf(point);
   return point;
 }
